Rename src to rprpp dir
Make directory correspond to target names
