+++ capi/rprpp.cpp
@@ -1,8 +1,8 @@
 #include "rprpp.h"
 
-#include "Error.h"
-#include "PostProcessing.h"
-#include "vk_helper.h"
+#include "rprpp/Error.h"
+#include "rprpp/PostProcessing.h"
+#include "rprpp/vk_helper.h"
 
 #include <cassert>
 #include <concepts>
