+++ rprpp/DescriptorBuilder.cpp
