+++ rprpp/Error.cpp
