+++ rprpp/ImageFormat.h
