+++ rprpp/common.h
