+++ rprpp/vk_helper.cpp
